Correct Queue::peek_ build errors on AVR-GCC 9.1.0
In addition to one overload passing a `T(*)[SIZE]` to `peek_(T*, uint8_t)`, the
queue peeking also uses a `min` macro from `math.h` which is non-standard
and not present in the AVR-LibC 2.0 (as far as I can tell). This commit fixes
the pointer type issue and replaces the `min` macro with a ternary expression.