Fix potential overflow in raw value conversion utilities
The utility functions for mapping raw/physical values suffer from
integer overflow and issues with conversions between signed and
unsigned values. In order to handle values near the edge of the
input ranges, calculations must be done in a larger type (i.e.
int32_t when the input is int16_t).